[upgrade] Make 'apple-bce' upgrade idempotent
Related to this, we now update the 'VERSION' file to have the latest
version with a suffix of '-UPGRADED' to denote that it was an upgrade
and not a fresh installation.

Resolves #73

+++ scripts/upgrade-arch-linux.sh
@@ -48,10 +48,15 @@ fi
 echo "Upgrading Linux kernels by adding Linux LTS 5.4 complete."
 
 echo "Upgrading Mac drivers..."
-dkms remove -m apple-bce -v 0.1 --all
-rm -rf /usr/src/apple-bce-0.1
-git clone https://github.com/ekultails/mbp2018-bridge-drv --branch mac-linux-gaming-stick /usr/src/apple-bce-0.1
-dkms install -m apple-bce -v 0.1 -k $(ls -1 /usr/lib/modules/ | grep -P "^[0-9]+")
+if [[ "$(cat /etc/mac-linux-gaming-stick/VERSION)" == "2.0.0" ]];
+    then echo "Installing new 'apple-bce' driver..."
+    dkms remove -m apple-bce -v 0.1 --all
+    rm -rf /usr/src/apple-bce-0.1
+    git clone https://github.com/ekultails/mbp2018-bridge-drv --branch mac-linux-gaming-stick /usr/src/apple-bce-0.1
+    dkms install -m apple-bce -v 0.1 -k $(ls -1 /usr/lib/modules/ | grep -P "^[0-9]+")
+else
+    echo "Skipping installing 'apple-bce' (Mac Linux Gaming Stick '2.0.0' detected)."
+fi
 echo "Upgrading Mac drivers complete."
 
 echo "Upgrading GRUB menu..."
@@ -112,5 +117,7 @@ fi
 chmod +x /home/stick/Desktop/*.desktop
 chown -R stick: /home/stick/Desktop/*.desktop
 echo "Upgrading desktop shortcuts complete."
+
+echo "2.1.0-UPGRADED" > /etc/mac-linux-gaming-stick/VERSION
 
 echo "Running 2.0.0 to 2.1.0 upgrades complete."
